infosec/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
cryptography==40.0.1
pycryptodome==3.17.0
qrcode[pil]==7.4.2
pyzbar==0.1.9
Pillow==9.5.0
python-dotenv==1.0.0